% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ontogeny_param}
\alias{ontogeny_param}
\title{Human metabolic enzyme and drug transporter ontogeny data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 89 rows and 13 columns.
}
\source{
<https://doi.org/10.1007/s40262-014-0140-7>

<https://doi.org/10.1002/jcph.585>

<https://doi.org/10.1002/cpt.1516)>
}
\usage{
ontogeny_param
}
\description{
Retrieved from the literature, see 'source':
}
\keyword{datasets}
