% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontogeny_functions.R
\name{ontogeny_plot}
\alias{ontogeny_plot}
\title{Plot enzyme or transporter activity over age}
\usage{
ontogeny_plot(name, loc = "hepatic", min_age = 0, max_age = 4, age_step = 0.02)
}
\arguments{
\item{name}{The target enzyme or transporter as character.}

\item{loc}{the localisation as character. Can be 'hepatic', 'intestinal',
'brain' or 'renal'.}

\item{min_age}{Minimum age in years as numeric.}

\item{max_age}{Maximum age in years as numeric.}

\item{age_step}{The age resolution for the plot in years, as numeric.}
}
\value{
A ggplot object.
}
\description{
Plot enzyme or transporter activity over age
}
\examples{
ontogeny_plot("CYP3A4")
ontogeny_plot(c("CYP3A4", "CYP2C9", "CYP2C19"))
ontogeny_plot(c("UGT1A1", "UGT1A4", "UGT1A6"), max_age = 6)
}
