% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontogeny_functions.R
\name{rel_act}
\alias{rel_act}
\title{Relative enzyme or transporter activity by age}
\usage{
rel_act(name, age, loc = "hepatic")
}
\arguments{
\item{name}{The target enzyme or transporter as character.}

\item{age}{The age as numeric.}

\item{loc}{The localisation as character. Can be 'hepatic', 'intestinal',
'brain' or 'renal'.}
}
\value{
The relative activity as numeric.
}
\description{
Relative enzyme or transporter activity by age
}
\examples{
rel_act("CYP3A4", 1)
rel_act(c("CYP3A4", "CYP2D6"), 1)
}
